open-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opencli"
version = "0.1.0"
description = "AI-assisted programming CLI tool"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["cli", "ai", "programming", "assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.25.0",
    "pyyaml>=6.0",
    "markdown>=3.0",
    "Pygments>=2.0",
    "blessed>=1.0",
    "pymupdf>=1.0",
    "requests>=2.0",
    "beautifulsoup4>=4.0",
    "pillow>=10.0",
    "pytesseract>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
opencli = "cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/open-cli"
Repository = "https://github.com/yourusername/open-cli"

[tool.setuptools]
py-modules = ["cli", "config"]
packages = ["core", "tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
